ft: improvements in code quality - batch 1
